# backend/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3001",
  "basePath": "/",
  "tags": [
    {
      "name": "NFT",
      "description": "與合約、NFT 相關的 api"
    },
    {
      "name": "Firebase",
      "description": "與 Firebase 內合約資訊做交互的 api "
    },
    {
      "name": "Pinata",
      "description": "與上傳 metadata.json 相關的 api"
    },
    {
      "name": "User",
      "description": "與 Firebase 內用戶資訊做交互的 api "
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/nft/mint": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "contractAddress": {
                  "example": "any"
                },
                "receiver": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/nft/deploy": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "abbreviation": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "imageUrl": {
                  "example": "any"
                },
                "metadataUrls": {
                  "example": "any"
                },
                "holder": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/firebase/read/{address}": {
      "get": {
        "tags": [
          "Firebase"
        ],
        "description": "",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/firebase/store": {
      "post": {
        "tags": [
          "Firebase"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "contractAddress": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "imageUrl": {
                  "example": "any"
                },
                "metadataUrls": {
                  "example": "any"
                },
                "holder": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/firebase/contracts": {
      "get": {
        "tags": [
          "Firebase"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/pinata/upload_json": {
      "post": {
        "tags": [
          "Pinata"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "jsonfile": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/pinata/get_metadata/{qid}": {
      "get": {
        "tags": [
          "Pinata"
        ],
        "description": "",
        "parameters": [
          {
            "name": "qid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/mycontract/{useraddress}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "useraddress",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/nfts/{useraddress}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "useraddress",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/mycontract/join/{useraddress}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "useraddress",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}